refactor(projects-service): Move storage deletion to ContainerHandler
